#11 Move activation condition processing to MainProcessor, make EEL work
It makes more sense to have this in MainProcessor because this should
contain the majority of processing logic and be able to work on data
structures that are suitable for processing. It became obvious when
realizing that - in order to implement EEL activation conditions -
in the build stage (Session) we need a string but no EEL processor,
in the processing stage (MainProcessor) we need an EEL processor
but no string.