#59 Fix crash by fixing threading issue
`RealearnPluginParameters::set_parameters()` is usually called by non-main thread!
That's why session access in this method is actually a no-go. Now
`RealearnPluginParameters` is the canonical owner of the parameter array and syncs
it directly to the processor via channel. The session doesn't know the parameter
values anymore (it doesn't need to know them). `ActivationCondition` API was
adjusted accordingly. In addition to having solved the race condition, things should
be much better decoupled now.