#84 Detect play state change via REAPER function instead of VST host callback
Because querying the play state needs a project and in ReaLearn we never just take
the currently active project, we also needed to make the real-time processor aware
of the containing project. Because the containing project is not immediately
available at plug-in instantiation time, we needed to use the Rc<Lazy<...>>
technique that's also used with Session.